Добавлен Slow Link.
Добавлен интерфейс Slow Link, который реализован на базе класса MCHS_Imitator.

+++ Controller/binp_device.h
@@ -32,18 +32,18 @@ public:
 
     int getStartAddrOfDownLink(int numOfDownLink);
     // - - - Функции, использующие 130 и 137 команды. Используются всеми платами, использующие C-Link - - - - - - - - - - - - - //
-    bool USBtoCLink_RedirRead(int numOfDownLink, unsigned short CL_Addr, QList<unsigned short> *Data);  // чтение 62-x слов КЛ
-    bool CLink_RedirWrite(int numOfDownLink, unsigned short CL_Addr, QList<unsigned short>* Data);  // запись данных в КЛ (8 слов)
-    bool CLink_TxRx(int numOfDownLink, int opcode, QList<unsigned short>* Addr, QList<unsigned short>* Data);  // запись данных в КЛ (любого кол-во слов)
+    bool USBtoCLink_RedirRead(int numOfDownLink, unsigned int CL_Addr, QList<unsigned int> *Data);  // чтение 62-x слов КЛ
+    bool CLink_RedirWrite(int numOfDownLink, unsigned int CL_Addr, QList<unsigned int>* Data);  // запись данных в КЛ (8 слов)
+    bool CLink_TxRx(int numOfDownLink, int opcode, QList<unsigned int>* Addr, QList<unsigned int>* Data);  // запись данных в КЛ (любого кол-во слов)
     bool CLink_sendOpcode(int numOfDownLink,int opcode);
     // - - - Функции, использующие 145 и 154 команды. Разработка Юдина Ю.В. - - - - - - - - - - - - - - - - - - - - - - - - - - //
-    bool SendWrite4  (int numOfDownLink, int N, QList<unsigned short>* Addr, QList<unsigned short>* Data);  // запись данных (до 4 слов)
-    bool SendWrite154(int numOfDownLink, int N, QList<unsigned short>* Addr, QList<unsigned short>* Data);  // запись любого кол-ва данных
-    bool SendRead8   (int numOfDownLink, int N, QList<unsigned short>* Addr, QList<unsigned short> *Data);  // чтение данных (до 8 слов)
-    bool SendRead145 (int numOfDownLink, int N, QList<unsigned short>* Addr, QList<unsigned short> *Data);  // чтение любого кол-ва данных
+    bool SendWrite4  (int numOfDownLink, int N, QList<unsigned int>* Addr, QList<unsigned int>* Data);  // запись данных (до 4 слов)
+    bool SendWrite154(int numOfDownLink, int N, QList<unsigned int>* Addr, QList<unsigned int>* Data);  // запись любого кол-ва данных
+    bool SendRead8   (int numOfDownLink, int N, QList<unsigned int>* Addr, QList<unsigned int> *Data);  // чтение данных (до 8 слов)
+    bool SendRead145 (int numOfDownLink, int N, QList<unsigned int>* Addr, QList<unsigned int> *Data);  // чтение любого кол-ва данных
 
     bool Init_Waiting193(int numOfDownLink); // Инициальзация МЧС для ожижания 193 команды
-    bool ReadDataFromDownLink (int numOfDownLink, QList<unsigned short> *Data);  // чтение данных по 193 команде
+    bool ReadDataFromDownLink (int numOfDownLink, QList<unsigned int> *Data);  // чтение данных по 193 команде
 };
 
 class ADIS
